moved all card manipulation functions to card_tools
